backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
spacy
sentence-transformers
pydantic
python-docx
